backend/requirements.txt
alembic==1.10.2
anyio==3.6.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
dnspython==2.3.0
email-validator==1.3.1
Faker==18.3.1
fastapi==0.95.0
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
pydantic==1.10.7
PyJWT==2.6.0
pytest==7.2.2
python-dateutil==2.8.2
python-multipart==0.0.6
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.20.0
prometheus-fastapi-instrumentator==6.0.0
psycopg2-binary==2.9.6
